refactor(ts): break the cli format-all <-> composition-root import cycle
Extract FormatAllCommand into cli/format-all-command.ts (importing only its
direct deps), matching format-pass-command.ts/validate-syntax-command.ts.
cli/format-all.ts is now a pure entry (main + run-as-main guard importing only
CompositionRoot); composition-root.ts imports the command module. sidecar.ts
still imports #sidecar/cli/format-all.
